requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Image Upload
cloudinary==1.36.0
python-magic==0.4.27
pillow==11.3.0

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Utils
python-dotenv==1.0.0
